feat: STABLE dev deployment

+++ dev/utils/log_utils.py
@@ -2,7 +2,8 @@
 import os
 import shutil
 
-from settings import config, PROJECT_ROOT_PATH
+from dev.utils.base_utils import get_tmp_dir_path
+from settings import config
 
 
 def use_tmp_dir_for_logs(need_to_clear_logs: bool = True):
@@ -16,14 +17,11 @@ def use_tmp_dir_for_logs(need_to_clear_logs: bool = True):
         - new path definition for logging (PROJECT_DIR/tmp/logs/)
      """
     assert config.deploy.mode == 'development'
-    tmp_dir_path = PROJECT_ROOT_PATH / 'tmp'
+    tmp_dir_path = get_tmp_dir_path()
     log_dir = tmp_dir_path / 'logs'
 
-    if not os.path.exists(tmp_dir_path):
-        os.mkdir(tmp_dir_path)
-    else:
-        if need_to_clear_logs:
-            shutil.rmtree(log_dir)
+    if os.path.exists(log_dir) and need_to_clear_logs:
+        shutil.rmtree(log_dir)
 
     if not os.path.exists(log_dir):
         os.mkdir(log_dir)
